- fix: gameId 타입 변경(Int→Long)에 따른 타입 불일치 오류 수정 (MemberGameAddRequest, MemberGameDto, ReviewWriteRequest, ReviewDto, MemberGameRepository MemberGameService,ApiV1ReviewController) 등  - ReviewService, MemberGameService: game.getId() → game.id, requireNotNull() 적용 - 테스트: addGameToLibrary(gameId) int → long, setReview() → 프로퍼티 직접 할당

+++ src/main/kotlin/com/back/domain/member/memberGame/dto/MemberGameAddRequest.kt
@@ -9,5 +9,5 @@ data class MemberGameAddRequest(
     @field:DecimalMin("0.0") val playtime: Double,
     val isFavorite: Boolean,
     val status: StatusEnum?,
-    val gameId: Int,
+    val gameId: Long,
 )
